deps: downgrade transformers to <5 and port deepseek_v2 backend
Pin transformers>=4.45,<5, huggingface_hub>=0.24,<1, and add
tokenizers>=0.20,<0.22 (previously transformers==5.4.0 /
huggingface_hub==1.10.1).

transformers 5.x had refactored the deepseek_v2 MoE into a fused
`DeepseekV2Experts` module and renamed the block `DeepseekV2Moe`; neither
exists on 4.x (experts there are an nn.ModuleList of MLPs). Port the
active backend so it no longer depends on those v5-only symbols:

- custom_deepseek_v2_lite.py: reparent CustomDeepseekV2Experts and
  CustomDeepseekV2Moe to nn.Module (keeping our fused 3D-tensor expert
  design and its custom fused<->per-expert state-dict serialization, which
  defines the on-chain/HF checkpoint format), and add an explicit
  CustomDeepseekV2Moe.forward that replicates v5's inherited MoE forward.
- mycelia.py: pass torch_dtype= (not dtype=) to from_pretrained; 4.x has
  no dtype alias, so dtype= was silently ignored and the model loaded fp32.

Verified under an installed transformers 4.55.4: clean imports, full +
partial model forward, state-dict round-trip, and bit-identical logits
(max|delta|=0) vs 5.4.0 with the same weights and input.

+++ connito/shared/modeling/mycelia.py
@@ -110,7 +110,10 @@ def load_pretrained_model_low_mem(
     model = model_class.from_pretrained(
         model_path,
         config=moe_config,
-        dtype=model_dtype,
+        # `torch_dtype` (not `dtype`): transformers 4.x has no `dtype` alias for
+        # `from_pretrained`, so passing `dtype` there is silently ignored and the
+        # model loads at the default fp32 — doubling host RAM on this low-mem path.
+        torch_dtype=model_dtype,
         low_cpu_mem_usage=False,
         ignore_mismatched_sizes=True,
     )
